% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_jitter_to_mean_pick.R
\name{add_jitter_to_mean_pick}
\alias{add_jitter_to_mean_pick}
\title{Add Jitter to Mean Pick Values}
\usage{
add_jitter_to_mean_pick(
  Final.working.data,
  mean.pick,
  means,
  sds,
  verbose = TRUE
)
}
\arguments{
\item{Final.working.data}{A numeric matrix containing the normalized counts data
where the mean pick values will be modified.}

\item{mean.pick}{A numeric value indicating the mean threshold below which values
will be modified.}

\item{means}{A list of numeric vectors where each vector contains the means for each
group.}

\item{sds}{A list of numeric vectors where each vector contains the standard deviations
corresponding to each group.}
}
\value{
A numeric matrix with modified values where the mean pick has been jittered.
The output matrix maintains the same dimensions as the input matrix.
}
\description{
This function adds jitter to the values in the provided data matrix where the values
equal a specified mean pick. The jitter is generated based on the standard deviation
of the means corresponding to the groups. This is useful for softening values that
are exactly equal to the mean pick, helping to mitigate potential issues with
downstream analysis.
}
\details{
The function identifies the locations in the input matrix where the values are equal to
the mean pick, calculates the standard deviations associated with these mean values,
generates jitter based on these standard deviations, and replaces the mean pick values
with the generated jitter. If no corresponding standard deviations are found, a warning
message is displayed, and the function proceeds without jittering the mean pick values.
}
\keyword{internal}
