% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_conversion.R
\name{apply_conversion}
\alias{apply_conversion}
\title{Apply Binary Conversion to Categorical Covariates}
\usage{
apply_conversion(df, cols, verbose = T)
}
\arguments{
\item{df}{A data frame containing the covariates to be transformed. This data frame should consist of
categorical columns specified in the \code{cols} parameter.}

\item{cols}{A character vector of column names to be converted. These should be categorical columns
in the data frame \code{df}.}

\item{verbose}{A logical value indicating whether to print messages about the transformation process
to the console. Defaults to TRUE.}
}
\value{
A data frame with the specified columns converted to binary format. A message is printed to
the console listing all the columns that were transformed or notifying if a specified column was not found.
}
\description{
This internal function applies binary conversion to specified categorical columns in a DataFrame.
For factors with two levels, it converts them to binary (0/1). For factors with more than two levels,
it creates new binary columns for each level and removes the original column.
}
\keyword{internal}
