% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cv.0.R
\name{calculate_cv}
\alias{calculate_cv}
\title{Calculate Coefficient of Variation
This function calculates the Coefficient of Variation (CV) based on a provided mean and LOESS smoothing results.
It identifies the standard deviations corresponding to a specific mean from the LOESS results
and computes the CV as twice the mean of these standard deviations.}
\usage{
calculate_cv(mean.pick, lowess)
}
\arguments{
\item{mean.pick}{A numeric value representing the target mean for which to calculate the CV.}

\item{lowess}{A list containing LOESS results, where each entry includes 'x' (means) and 'y' (standard deviations).}
}
\value{
A numeric value representing the Coefficient of Variation (CV) calculated as twice the mean of the selected standard deviations.
If no valid indices are found, it returns a message indicating so.
}
\description{
Calculate Coefficient of Variation
This function calculates the Coefficient of Variation (CV) based on a provided mean and LOESS smoothing results.
It identifies the standard deviations corresponding to a specific mean from the LOESS results
and computes the CV as twice the mean of these standard deviations.
}
\keyword{internal}
