% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_categorical_factors.R
\name{check_categorical_factors}
\alias{check_categorical_factors}
\title{Check and Convert Categorical Variables to Factors}
\usage{
check_categorical_factors(data_frame, variables)
}
\arguments{
\item{data_frame}{A data frame containing the variables to be checked and converted.}

\item{variables}{A character vector of variable names in \code{data_frame} that are intended to be categorical.}
}
\value{
The original data frame with specified categorical variables converted to factors if they were not already.
}
\description{
This function checks specified variables in a data frame to ensure they are factors.
If any specified categorical variables are not factors, they are converted to factors.
}
\details{
The function identifies columns in the data frame that are categorical (non-numeric). It then checks
whether these categorical columns are factors. If any of the specified categorical variables are not factors,
they are converted to factors, and a message is displayed listing the converted variables. If all specified
categorical variables are already factors, a message indicating this is printed.
}
\keyword{internal}
