% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_contrast_matrix.R
\name{create_contrast_matrix}
\alias{create_contrast_matrix}
\title{Create Contrast Matrix for Falcon Package Analysis}
\usage{
create_contrast_matrix(fit, contrasts, reference.level, n.levels)
}
\arguments{
\item{fit}{An object of class \code{lm} from a model fit. This object is used to extract
the terms from the model matrix to construct the contrast matrix.}

\item{contrasts}{A character vector specifying the contrasts to be created. Each contrast should be
defined as a comparison between two levels of a factor, separated by a colon (e.g., "level1:level2").
The levels should correspond to the terms in the model matrix.}

\item{reference.level}{A character string indicating the reference level for contrasts. Levels specified
as the reference level will be compared against other levels in the contrasts.}

\item{n.levels}{A character vector containing the levels of the categorical variable in the model.
This vector is used to locate the relevant columns for contrast vectors.}
}
\value{
A matrix where each row represents a specified contrast. The columns correspond to the model
terms, and each row contains values of 1, -1, and 0 that define the contrast vector for comparing
the specified levels.
}
\description{
Generates a contrast matrix for use in differential expression analysis within the Falcon package.
The contrast matrix is designed to facilitate hypothesis testing between specified levels
of a categorical variable in a linear model fit.
}
\details{
This function is used within the Falcon package to construct contrast matrices for differential expression
analysis. The process includes:
\itemize{
\item Initializing an empty list to store contrast vectors.
\item Iterating through the provided contrast specifications.
\item Splitting each contrast into two levels.
\item Creating and populating a contrast vector with values corresponding to the levels in the model matrix.
\item Combining all individual contrast vectors into a single matrix.
\item Setting row names of the matrix to the contrast specifications.
}
}
\keyword{internal}
