% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FalconNest.R
\name{dropCovariate}
\alias{dropCovariate}
\title{Drop Covariate}
\usage{
dropCovariate(object, covariate)
}
\arguments{
\item{object}{An instance of the \code{FalconNest} class.}

\item{covariate}{A vector containing one or more covariates to be dropped.}
}
\description{
Retrieves the counts data from the \code{FalconNest} object based on the specified type.
See \code{Falcon}, \code{FalconPreen}, \code{FalconNest}.
}
\examples{
# Example data
colData <- data.frame(sample = factor(rep(c("A", "B"), each = 5)),
cov1 = factor(rep(c("C", "D"), each = 5)))
counts <- matrix(data = abs(rpois(50 * 10, lambda = 5)), nrow = 50, ncol = 10)
colnames(counts) <- rownames(colData)
rownames(counts) <- 1:50

# Initialize FalconNest object
falcon_obj <- FalconNest(counts = counts, colData = colData, design = ~sample+cov1)

# Drop covariate
falcon_obj <- dropCovariate(object = falcon_obj, covariate = "cov1")

# Perform analysis
falconResults <- FalconHatch(object = normalizedObject, random.covar = FALSE, shuffle.group = FALSE, contrasts = "all", FalconOutput= FALSE)
adjExpression <- getExpression(object = falconResults, type = "adjusted")

}
