% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runANCOVA.R
\name{FalconGlide}
\alias{FalconGlide}
\title{Falcon Function for ANCOVA Analysis}
\arguments{
\item{object}{An object of class \code{FalconNest}, containing counts data, sample metadata, and normalized counts.}

\item{contrasts}{A character string specifying the type of contrasts for post-hoc
testing. Defaults to "all" for Tukey's HSD test. Other options include "reference" for Dunnett's test or custom contrasts.}

\item{random.covar}{Logical, indicating whether to apply random covariate control. Defaults to TRUE.}

\item{shuffle.group}{Logical, indicating whether to shuffle the group variable for control. Defaults to TRUE.}

\item{FalconOutput}{Logical, whether to capture the input parameters. Defaults to TRUE.}
}
\value{
An updated \code{FalconNest} object with two additional slots:
\item{AdjustedCounts}{A data frame of adjusted counts after ANCOVA.}
\item{Results}{A data frame of ANCOVA results, including post-hoc analysis, if applicable.}
}
\description{
This function performs ANCOVA analysis on the normalized counts data from a
\code{FalconNest} object.
}
\details{
This function runs ANCOVA on each gene in the \code{NormalizedCounts} slot of the \code{FalconNest} object.
It builds the ANCOVA model based on the provided design and group variables, with options for
post-hoc analysis using Tukey's HSD, Dunnett's test, or custom contrasts. The function also
adjusts for multiple comparisons using the Benjamini-Hochberg method and can perform controls
for random covariates or shuffled group variables.
}
\keyword{internal}
