% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FalconNest.R
\name{FalconHatch}
\alias{FalconHatch}
\title{Falcon differential analysis with covariate selection}
\usage{
FalconHatch(object, ...)
}
\arguments{
\item{object}{An instance of the \code{FalconNest} class.}

\item{contrasts}{A character string or matrix specifying the contrasts for post-hoc testing.
Default is "all".}
}
\value{
The updated \code{FalconNest} object with results from the FalconHatch analysis.
}
\description{
Applies ANCOVA analysis on the \code{FalconNest} object and updates the results.
The analysis uses the normalized counts and the specified contrasts.
}
\examples{
# Create example data
colData <- data.frame(sample = factor(rep(c("A", "B"), each = 5)))
counts <- matrix(data = abs(rpois(50 * 10, lambda = 5)), nrow = 50, ncol = 10)
colnames(counts) <- rownames(colData)
rownames(counts) <- 1:50
falcon_obj <- FalconNest(counts = counts, colData = colData, design = ~sample)
normalizedObject <- FalconPreen(object = falcon_obj, group = "sample", mean.pick = 1, cv.pick = 100)
# Perform analysis
falconResults <- FalconHatch(object = normalizedObject, random.covar = FALSE, shuffle.group = FALSE, contrasts = "all", FalconOutput= FALSE)
}
