% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FalconHoming.R
\name{FalconHoming}
\alias{FalconHoming}
\title{FalconHoming: Validation and reproducibility of Falcon analysis.}
\usage{
FalconHoming(counts, colData = NULL, FalconOutput = NULL)
}
\arguments{
\item{counts}{A numeric matrix of gene expression counts where rows are genes
and columns are samples.}

\item{colData}{A DataFrame containing sample metadata}

\item{FalconOutput}{A character string indicating a path to a
directory where the '.Falcon' is stored. If not provided,
the function will search for existing \code{.Falcon} files in the current directory.}
}
\value{
A FalconHoming object containing results of the analysis.
}
\description{
This function conducts all the basic steps of Falcon pipeline from data
normalization to differential expression analysis using FalconOutput log file.
}
\examples{
# Example usage of FalconHoming
result <- FalconHoming(counts = my_counts, colData = my_colData)

}
