% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FalconNest.R
\docType{class}
\name{FalconNest-class}
\alias{FalconNest-class}
\title{FalconNest Class Definition}
\description{
An S4 class to represent the \code{FalconNest} object, which encapsulates
counts data, sample metadata, and results from various analyses.
The class also includes internal slots for design formulas and
normalization data.
}
\section{Slots}{

\describe{
\item{\code{counts}}{A matrix or data frame containing counts data.}

\item{\code{colData}}{A data frame with sample metadata.}

\item{\code{NormalizedCounts}}{A matrix or data frame of normalized counts.}

\item{\code{AdjustedCounts}}{A matrix or data frame of adjusted counts.}

\item{\code{Results}}{A matrix or data frame containing results from analysis.}

\item{\code{.design}}{A formula representing the experimental design (internal slot).}

\item{\code{.GroupVariable}}{A character string for the group variable used in analysis (internal slot).}

\item{\code{.InternalNormList}}{A list containing internal normalization data (internal slot).}

\item{\code{.FalconHatch}}{A list for internal use (internal slot).}
}}

