% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FalconNest.R
\name{FalconNest}
\alias{FalconNest}
\title{Constructor for FalconNest Class}
\usage{
FalconNest(counts, colData, design)
}
\arguments{
\item{counts}{A matrix or data frame of counts data.}

\item{colData}{A data frame containing sample metadata.}

\item{design}{A formula representing the experimental design.}
}
\value{
An instance of the \code{FalconNest} class.
}
\description{
Initializes a new \code{FalconNest} object with counts data, sample metadata, and an experimental design formula.
Internal slots are set to default values, and the object is ready for further analysis and processing.
}
\examples{
# Create example data
colData <- data.frame(sample = factor(rep(c("A", "B"), each = 5)))
counts <- matrix(data = abs(rpois(50 * 10, lambda = 5)), nrow = 50, ncol = 10)
colnames(counts) <- rownames(colData)
rownames(counts) <- 1:50
falcon_obj <- FalconNest(counts = counts, colData = colData, design = ~sample)
}
