% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FalconScout.R
\name{FalconScout}
\alias{FalconScout}
\title{FalconScout: Visualize PCA and Density Plots for Informed Covariate Selection}
\usage{
FalconScout(object, group)
}
\arguments{
\item{object}{A FalconNest class object.}

\item{group}{A character string specifying the grouping variable in colData to be used for plotting.}
}
\value{
A grid of PCA and density plots organized by the variables in the design.
}
\description{
This function generates PCA plots for categorical variables and density plots for continuous variables
from a FalconNest object. It normalizes the counts and organizes plots
based on the specified group variable passed to FalconNest().
}
