% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FalconNest.R
\name{getExpression}
\alias{getExpression}
\title{Retrieve expression values}
\usage{
getExpression(object, type)
}
\arguments{
\item{object}{An instance of the \code{FalconNest} class.}

\item{type}{A character string specifying the type of data to retrieve. Options are "raw", "normalized", or "adjusted".}
}
\value{
The requested type of counts data.
}
\description{
Retrieves the counts data from the \code{FalconNest} object based on the specified type.
See \code{Falcon}, \code{FalconPreen}, \code{FalconNest}.
}
\examples{
# Example data
colData <- data.frame(sample = factor(rep(c("A", "B"), each = 5)))
counts <- matrix(data = abs(rpois(50 * 10, lambda = 5)), nrow = 50, ncol = 10)
colnames(counts) <- rownames(colData)
rownames(counts) <- 1:50

# Initialize FalconNest object
falcon_obj <- FalconNest(counts = counts, colData = colData, design = ~sample)

# Normalize counts
normalizedObject <- FalconPreen(object = falcon_obj, group = "sample", mean.pick = 1, cv.pick = 100)

# Perform analysis
falconResults <- FalconHatch(object = normalizedObject, random.covar = FALSE, shuffle.group = FALSE, contrasts = "all", FalconOutput= FALSE)
adjExpression <- getExpression(object = falconResults, type = "adjusted")
}
