% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FalconNest.R
\name{getResults}
\alias{getResults}
\title{Extract posthoc group comparisons}
\usage{
getResults(object, p.adjusted = 0.05, extendedData = FALSE)
}
\arguments{
\item{object}{An instance of the \code{FalconNest} class.}

\item{p.adjusted}{A numeric value specifying the threshold for adjusted p-values (default is 0.05).}

\item{extendedData}{A logical value indicating whether to return a detailed results table,
including degrees of freedom, confidence interval, and unadjusted expression (default is FALSE).}
}
\value{
A data frame of results that meet the specified criteria.
}
\description{
Retrieves the results from the Falcon analysis based on specified p-value adjustment threshold and data type.
See \code{Falcon}, \code{FalconPreen}, \code{FalconNest}.
}
\examples{
# Example data
colData <- data.frame(sample = factor(rep(c("A", "B"), each = 5)))
counts <- matrix(data = abs(rpois(50 * 10, lambda = 5)), nrow = 50, ncol = 10)
colnames(counts) <- rownames(colData)
rownames(counts) <- 1:50

# Initialize FalconNest object
falcon_obj <- FalconNest(counts = counts, colData = colData, design = ~sample)

# Normalize counts
normalizedObject <- FalconPreen(object = falcon_obj, group = "sample", mean.pick = 1, cv.pick = 100)

# Perform analysis
falconResults <- FalconHatch(object = normalizedObject, random.covar = FALSE, shuffle.group = FALSE, contrasts = "all", FalconOutput= FALSE)
falconResults <- getResults(object = falconResults, p.adjusted = 0.05, extendedData = FALSE)

# Print results
print(falconResults)
}
