% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FalconNest.R
\docType{class}
\name{MatrixOrDataFrame-class}
\alias{MatrixOrDataFrame-class}
\title{MatrixOrDataFrame Class Union}
\description{
Defines a class union for objects that are either
matrices or data frames. This allows functions
and methods to handle both types of objects interchangeably.
}
