% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normData.R
\name{normData}
\alias{normData}
\title{R/normData.R
Normalize Counts and Perform Quality Checks}
\usage{
normData(
  object,
  group,
  mean.pick = NULL,
  cv.pick = NULL,
  flooring.option = "soft",
  flooring.ratio = 0.6,
  lowess.f = 1/64
)
}
\arguments{
\item{object}{A \code{FalconNest} object containing count data, metadata, and design information.}

\item{group}{The name of the grouping variable in the \code{colData} slot of the \code{FalconNest} object, which will be used for normalization and
filtering.}

\item{mean.pick}{An optional threshold for mean values. If \code{NULL}, the function will compute this threshold based on combined data from all groups.}

\item{cv.pick}{A numeric value representing the maximum acceptable coefficient of variation (CV) for filtering. Default is 20.}

\item{flooring.option}{A character string specifying the flooring method: "soft" or "hard".}

\item{flooring.ratio}{A numeric value between 0 and 1 used in soft flooring to determine the ratio of samples above \code{mean.pick}.}

\item{lowess.f}{A numeric value representing the span for LOWESS smoothing.}
}
\value{
The updated \code{FalconNest} object with normalized counts and additional internal results stored.
}
\description{
This function normalizes count data from a \code{FalconNest} object using a specified normalization method. It performs various quality checks
and filtering based on mean and coefficient of variation (CV) thresholds. The function also updates the \code{FalconNest} object with the
normalized data and stores internal results for further analysis.
}
\details{
\itemize{
\item The function first extracts count data and metadata from the \code{FalconNest} object and checks the design formula for required variables.
\item It normalizes the count data using counts per million (CPM) and log transformation, followed by CyclicLoess normalization between samples.
\item Quality checks are performed on the design formula, including checking the standard deviation of numeric variables.
\item The function applies binary transformation to factor covariates and calculates mean, standard deviation, CV, and performs LOWESS smoothing for each group.
\item If \code{mean.pick} is \code{NULL}, the function computes this threshold based on combined data from all groups. Otherwise, it uses the provided \code{mean.pick}.
\item Data is filtered based on the mean and CV thresholds, and the results are stored in the \code{FalconNest} object.
}
}
\keyword{internal}
