% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCovariates.R
\name{plotCovariates}
\alias{plotCovariates}
\title{Process and Visualize Covariate Counts}
\arguments{
\item{object}{A FalconNest object containing the \code{Results} data frame with the \code{Model} column.}
}
\value{
A grid plot with two visualizations of covariate selection frequency.

A grid plot with two visualizations of covariate selection frequency.
}
\description{
This function processes model covariates, calculates their occurrence percentage,
and visualizes the results using bar and line plots.

This function processes model covariates, calculates their occurrence percentage,
and visualizes the results using bar and line plots.
}
\section{Initialize FalconNest object}{
falcon_obj <- FalconNest(counts = counts, colData = colData, design = ~sample)
}

\section{Normalize counts}{
normalizedObject <- normCounts(object = falcon_obj,
group = "sample", mean.pick = 1, cv.pick = 100)
falconResults <- Falcon(object = normalizedObject, contrasts = "all")
plotCovariates(falconResults)
}

\examples{
colData <- data.frame(sample = factor(rep(c("A", "B"), each = 5)))
counts <- matrix(data = abs(rpois(50 * 10, lambda = 5)), nrow = 50, ncol = 10)
colnames(counts) <- rownames(colData)
rownames(counts) <- 1:50

# Initialize FalconNest object
falcon_obj <- FalconNest(counts = counts, colData = colData, design = ~sample)

# Normalize counts
normalizedObject <- normCounts(object = falcon_obj,
                                group = "sample", mean.pick = 1, cv.pick = 100)
falconResults <- Falcon(object = normalizedObject, contrasts = "all")
plotCovariates(falconResults)

}
