% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC.R
\name{plotQC}
\alias{plotQC}
\title{Plot Quality Check for Principal Components and Normalized Counts
This function generates quality check plots for an object of class \code{FalconNest}.
It produces two plots: one for Principal Component Analysis (PCA) and one for boxplot of normalized counts.
These plots help in assessing the distribution and variation in the data after noise correction.}
\arguments{
\item{object}{An object of class \code{FalconNest} containing normalized counts and other relevant data.}

\item{group}{A character string specifying the column in \code{colData} to be used for grouping in the plots.}
}
\value{
A grid of plots is displayed, including:
\itemize{
\item A PCA plot showing the distribution of samples along the first two principal components, colored by group.
\item A boxplot of normalized counts for each variable, colored by group.
}
The function does not return a value but displays the plots directly.
}
\description{
Plot Quality Check for Principal Components and Normalized Counts
This function generates quality check plots for an object of class \code{FalconNest}.
It produces two plots: one for Principal Component Analysis (PCA) and one for boxplot of normalized counts.
These plots help in assessing the distribution and variation in the data after noise correction.
}
\details{
The function first ensures that the input object is of class \code{FalconNest} and contains the required slots.
It then extracts the relevant data and checks if the specified group is present in the \code{colData} slot.
PCA is performed on the normalized counts to visualize the variation in the data, and a boxplot is created to show
the distribution of normalized counts across variables. Both plots use color palettes to differentiate groups and
are arranged side-by-side with a title indicating that they are quality check plots.
}
\examples{
colData <- data.frame(sample = factor(rep(c("A", "B"), each = 5)))
counts <- matrix(data = abs(rpois(50 * 10, lambda = 5)), nrow = 50, ncol = 10)
colnames(counts) <- rownames(colData)
rownames(counts) <- 1:50

# Initialize FalconNest object
falcon_obj <- FalconNest(counts = counts, colData = colData, design = ~sample)

# Normalize counts
normalizedObject <- normCounts(object = falcon_obj, group = "sample", mean.pick = 1, cv.pick = 100)
plotQC(normalizedObject, "sample")

}
