% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResults.R
\name{plotResults}
\alias{plotResults}
\title{Plot Results of Gene Expression Analysis}
\usage{
plotResults(object, group, p.adjusted)
}
\arguments{
\item{object}{An object of class 'FalconNest' containing the gene expression data, normalization information,
and results for analysis.}

\item{group}{A character string specifying the name of the grouping variable to be used for coloring the points
in the PCA plot and annotating the heatmap.}

\item{p.adjusted}{A numeric value indicating the adjusted p-value threshold for selecting significant genes.}
}
\value{
This function does not return a value. It generates and displays a PCA plot and a heatmap.
If the number of significant genes is less than the number of samples, it will display a warning message
and only show the heatmap.
}
\description{
This function generates PCA plots and heatmaps from gene expression data stored in a
FalconNest object. It visualizes the results based on the adjusted counts and group information.
}
\examples{
# Assuming 'falcon_nest_object' is a properly constructed FalconNest object
plotResults(falcon_nest_object, group = "Group", p.adjusted = 0.05)

}
