% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FalconNest.R
\name{show,FalconNest-method}
\alias{show,FalconNest-method}
\title{Show Method for FalconNest}
\usage{
\S4method{show}{FalconNest}(object)
}
\arguments{
\item{object}{An object of class \code{FalconNest}. The object should contain the following slots:
\describe{
\item{\code{counts}}{A matrix of counts data.}
\item{\code{colData}}{A data frame containing metadata about the samples.}
\item{\code{design}}{A formula specifying the design of the experiment.}
}}
}
\description{
Displays a summary of the \code{FalconNest} object. This method provides an overview
of the object's main components, including the dimensions of the counts matrix,
the design formula, and the number of samples in the metadata.
}
